+++ webpack.config.js
@@ -1,7 +1,9 @@
 var webpack = require('webpack');
 
 
-function makeConfig(target) {
+function makeConfig(target, minify) {
+  var targetExt = minify ? '.min.js' : '.js';
+
   return {
     entry: './src/index.js',
     hot: false,
@@ -9,14 +11,14 @@ function makeConfig(target) {
       libraryTarget: 'umd',
       library: 'mobxReact',
       path: __dirname,
-      filename: (function() {
+      filename: (function () {
         switch (target) {
-          case 'browser': return 'index.js';
-          case 'native': return 'native.js';
-          case 'custom': return 'custom.js';
+          case 'browser': return 'index' + targetExt;
+          case 'native': return 'native' + targetExt;
+          case 'custom': return 'custom' + targetExt;
           default: throw new Error('Unexpected target: ' + target);
         }
-      }())
+      } ())
     },
     resolve: {
       extensions: ['', '.js'],
@@ -42,7 +44,7 @@ function makeConfig(target) {
         commonjs2: 'react',
         amd: 'react'
       },
-      'react-dom': (function() {
+      'react-dom': (function () {
         if (target === 'browser') return {
           root: 'ReactDOM',
           commonjs: 'react-dom',
@@ -50,19 +52,29 @@ function makeConfig(target) {
           amd: 'react-dom'
         };
         return false; // not external, will use empty-module  alias
-      }()),
+      } ()),
       mobx: 'mobx'
     },
     plugins: [
       new webpack.DefinePlugin({
         __TARGET__: JSON.stringify(target),
       })
-    ]
+    ].concat(
+      minify
+        ? [new webpack.optimize.UglifyJsPlugin({
+          compressor: {
+            screw_ie8: true,
+            warnings: false
+          }
+        })]
+        : []
+    )
   };
 }
 
 module.exports = [
   makeConfig('browser'),
   makeConfig('native'),
-  makeConfig('custom')
+  makeConfig('custom'),
+  makeConfig('browser', true)
 ];